backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
cloudinary
certifi
gunicorn
werkzeug